Added count queries
Added query_handler. The get_count function will get the number
of revisions which fulfill a criteria.
It can be filtered by: user, tag, size or date

The date filtering still need some work, because the db does not
store the timestamps as timedate, which makes difficult to
get revs between 2 dates. I recommend to change the extractor to
store the timestamps in a timedate format